Implemented stashing and updated testing.
I was thinking about repeated regexes and nested calls to regexes
and decided that I needed to solve this problem outright. I did so
by adding a mechanism to stash and restore internal data, and to
manage match results with a stack system instead of a single variable.

I also updated the tests and, at the moment, everything passes. :-)