Substantial docs revision; swallowed Scrooge::Any into Scrooge::Quantified
After comparing Scrooge::ZWA's sane default handling of _apply, vs
Scrooge::Quantified's lack of any handling of _apply, I decided to absorb
Scrooge::Any into Scrooge::Quantified. That required a minor update to
t/21-ANY.t

I have also make huge strides with the reorganization (and general writing)
of the documentaiton for Scrooge, as reflected in the bulk of the
changes in this commit.