First commit in a rewrite to use match_info
When I first wrote Scrooge, I decided that the pattern object should store all
of the match details within special fields in the pattern itself. This led to
all kinds of headaches for recursive and/or repetitive patterns, and my attempt
to address those led to rediculous gymnastics in the code.

After some rumination, I have realized that it would be much better to have the
pattern objects build a distinct data structure with all of the pertinent match
details. I refer to that structure as match_info. The structure is allocated
once, during the prep stage, and each occurence of a given pattern leads to a
new allocation within the match_info data structure. As long as all cached
calculations and match information are stored in these data structures, all
kinds of recursion and repetition are permissible with zero effort on my part.